+++ lib/src/widgets/text_area.dart
@@ -5,6 +5,7 @@ import '../core/color.dart';
 import '../core/cursor.dart';
 import '../core/grapheme_metrics.dart';
 import '../core/input.dart';
+import '../core/mouse_cursor.dart';
 import '../foundation/text_editing_controller.dart';
 import '../foundation/text_selection.dart';
 import '../framework/build_context.dart';
@@ -318,6 +319,7 @@ class _TextAreaState extends State<TextArea>
           autofocus: widget.autofocus,
           onFocusChange: handleFocusChange,
           child: PointerListener(
+            mouseCursor: MouseCursor.text,
             onPointerDown: handlePointerDown,
             child: _TextAreaLeaf(
               layoutMetrics: _layoutMetrics,